Add files via upload
Data is provided by the Framingham hospital or can be obtained over Kaggle (https://www.kaggle.com/amanajmera1/framingham-heart-study-dataset) where a subset of the data set is stored (and used here). It is fast open-access which makes is suitable for data integration. The problem that the data can be taken off the website can be neutralized by getting the data over the study provider directly. This however takes an email and around 2 days (https://biolincc.nhlbi.nih.gov/studies/framcohort/).